+++ lib/widgets/ny_pullable.dart
@@ -2,34 +2,42 @@ import 'package:flutter/material.dart';
 import 'package:pull_to_refresh_flutter3/pull_to_refresh_flutter3.dart';
 
 /// The Pullable widget helps you refresh the content.
-class Pullable extends StatelessWidget {
-  Pullable({super.key, required this.child, this.onRefresh, this.headerStyle});
+class Pullable extends StatefulWidget {
+  const Pullable(
+      {super.key, required this.child, this.onRefresh, this.headerStyle});
 
   /// Classic Header
-  Pullable.classicHeader({super.key, required this.child, this.onRefresh})
+  const Pullable.classicHeader({super.key, required this.child, this.onRefresh})
       : headerStyle = "ClassicHeader";
 
   /// WaterDrop Header
-  Pullable.waterDropHeader({super.key, required this.child, this.onRefresh})
+  const Pullable.waterDropHeader(
+      {super.key, required this.child, this.onRefresh})
       : headerStyle = "WaterDropHeader";
 
   /// MaterialClassic Header
-  Pullable.materialClassicHeader(
+  const Pullable.materialClassicHeader(
       {super.key, required this.child, this.onRefresh})
       : headerStyle = "MaterialClassicHeader";
 
   /// WaterDropMaterial Header
-  Pullable.waterDropMaterialHeader(
+  const Pullable.waterDropMaterialHeader(
       {super.key, required this.child, this.onRefresh})
       : headerStyle = "WaterDropMaterialHeader";
 
   /// Bezier Header
-  Pullable.bezierHeader({super.key, required this.child, this.onRefresh})
+  const Pullable.bezierHeader({super.key, required this.child, this.onRefresh})
       : headerStyle = "BezierHeader";
 
   final Widget child;
   final Function()? onRefresh;
   final String? headerStyle;
+
+  @override
+  State<Pullable> createState() => _PullableState();
+}
+
+class _PullableState extends State<Pullable> {
   final RefreshController refreshController =
       RefreshController(initialRefresh: false);
 
@@ -41,24 +49,26 @@ class Pullable extends StatelessWidget {
       header: headerType(),
       controller: refreshController,
       onRefresh: () async {
-        if (onRefresh == null) {
+        if (widget.onRefresh == null) {
           refreshController.refreshCompleted(resetFooterState: true);
+          setState(() {});
           return;
         }
-        if (onRefresh is Future Function()) {
-          await onRefresh!();
+        if (widget.onRefresh is Future Function()) {
+          await widget.onRefresh!();
         } else {
-          onRefresh!();
+          widget.onRefresh!();
         }
         refreshController.refreshCompleted(resetFooterState: true);
+        setState(() {});
       },
-      child: child,
+      child: widget.child,
     );
   }
 
   /// Returns the header type
   Widget headerType() {
-    switch (headerStyle) {
+    switch (widget.headerStyle) {
       case "ClassicHeader":
         {
           return const ClassicHeader();